Separate extension package.json
