+++ .gitignore
@@ -1,5 +1,6 @@
 out
 dist
+extension/dist
 node_modules
 .vscode-test/
 npm-debug.log
