+++ extension/package.json
@@ -1,0 +1,118 @@
+{
+  "name": "avm-debug-vscode-extension",
+  "displayName": "AVM Debug",
+  "version": "1.0.0",
+  "publisher": "algorand-vscode",
+  "description": "Debug extension for developing AVM transactions and smart contracts.",
+  "author": {
+    "name": "Algorand, llc"
+  },
+  "license": "MIT",
+  "keywords": [
+    "multi-root ready"
+  ],
+  "engines": {
+    "vscode": "^1.80.0"
+  },
+  "categories": [
+    "Debuggers"
+  ],
+  "private": true,
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/algorand/avm-debugger.git"
+  },
+  "bugs": {
+    "url": "https://github.com/algorand/avm-debugger/issues"
+  },
+  "main": "dist/extension.js",
+  "browser": "dist/web-extension.js",
+  "activationEvents": [
+    "workspaceContains:**/*.teal",
+    "onDebug",
+    "onDebugResolve:avm",
+    "onDebugDynamicConfigurations:avm"
+  ],
+  "workspaceTrust": {
+    "request": "never"
+  },
+  "contributes": {
+    "languages": [
+      {
+        "id": "teal",
+        "extensions": [
+          ".teal"
+        ],
+        "configuration": "teal-language-configuration.json"
+      }
+    ],
+    "grammars": [
+      {
+        "language": "teal",
+        "scopeName": "source.teal",
+        "path": "syntaxes/teal.tmLanguage.json"
+      }
+    ],
+    "breakpoints": [
+      {
+        "language": "teal"
+      }
+    ],
+    "debuggers": [
+      {
+        "type": "avm",
+        "languages": [
+          "teal"
+        ],
+        "label": "AVM Debug",
+        "program": "../out/src/cli.js",
+        "runtime": "node",
+        "configurationAttributes": {
+          "launch": {
+            "properties": {
+              "simulateTraceFile": {
+                "type": "string",
+                "description": "Transaction group simulation response with execution trace.",
+                "default": "${workspaceFolder}/path/to/simulateTraceFile.json"
+              },
+              "programSourcesDescriptionFile": {
+                "type": "string",
+                "description": "Description file for sources of programs appearing in transaction group.",
+                "default": "${workspaceFolder}/path/to/programSourcesDescriptionFile.json"
+              },
+              "stopOnEntry": {
+                "type": "boolean",
+                "description": "Automatically stop after launch.",
+                "default": true
+              }
+            }
+          }
+        },
+        "initialConfigurations": [
+          {
+            "type": "avm",
+            "request": "launch",
+            "name": "Debug AVM Transactions",
+            "simulateTraceFile": "${workspaceFolder}/path/to/simulateTraceFile.json",
+            "programSourcesDescriptionFile": "${workspaceFolder}/path/to/programSourcesDescriptionFile.json",
+            "stopOnEntry": true
+          }
+        ],
+        "configurationSnippets": [
+          {
+            "label": "AVM Debug",
+            "description": "A new configuration for replaying and debugging a Algorand transactions.",
+            "body": {
+              "type": "avm",
+              "request": "launch",
+              "name": "Debug AVM Transactions",
+              "simulateTraceFile": "^\"\\${workspaceFolder}/path/to/simulateTraceFile.json\"",
+              "programSourcesDescriptionFile": "^\"\\${workspaceFolder}/path/to/programSourcesDescriptionFile.json\"",
+              "stopOnEntry": true
+            }
+          }
+        ]
+      }
+    ]
+  }
+}
