+++ extension/src/activateAvmDebug.ts
