+++ extension/src/configuration.ts
