+++ extension/src/internalDescriptorFactory.ts
@@ -2,7 +2,7 @@
 
 import * as vscode from 'vscode';
 import { ProviderResult } from 'vscode';
-import { AvmDebugSession } from './debugAdapter/debugRequestHandlers';
+import { AvmDebugSession } from '../../src';
 import { workspaceFileAccessor } from './fileAccessor';
 
 export class InlineDebugAdapterFactory
