+++ package.json
@@ -1,23 +1,14 @@
 {
   "name": "avm-debug",
-  "displayName": "AVM Debug",
   "version": "1.0.0",
-  "publisher": "algorand-vscode",
-  "description": "Debug extension for developing AVM transactions and smart contracts in VS Code.",
+  "description": "Algorand AVM transaction and smart contract debugger using the Debugger Adapter Protocol",
   "author": {
     "name": "Algorand, llc"
   },
   "license": "MIT",
-  "keywords": [
-    "multi-root ready"
-  ],
   "engines": {
-    "vscode": "^1.80.0"
+    "node": ">=18.0.0"
   },
-  "categories": [
-    "Debuggers"
-  ],
-  "private": true,
   "repository": {
     "type": "git",
     "url": "https://github.com/algorand/avm-debugger.git"
@@ -31,9 +22,9 @@
     "typecheck": "tsc -p tsconfig.json --noEmit",
     "check-format": "prettier . --check",
     "format": "prettier . --write",
-    "esbuild-base": "esbuild ./src/extension.ts --bundle --tsconfig=./tsconfig.json --external:vscode --format=cjs --platform=node --outfile=dist/extension.js",
+    "esbuild-base": "esbuild ./extension/src/extension.ts --bundle --tsconfig=./tsconfig.json --external:vscode --format=cjs --platform=node --outfile=extension/dist/extension.js",
     "watch": "npm run -S esbuild-base -- --sourcemap --sources-content=false --watch",
-    "esbuild-web": "esbuild ./src/web-extension.ts --bundle --tsconfig=./tsconfig.json --external:vscode --format=cjs --platform=browser --outfile=dist/web-extension.js",
+    "esbuild-web": "esbuild ./extension/src/web-extension.ts --bundle --tsconfig=./tsconfig.json --external:vscode --format=cjs --platform=browser --outfile=extension/dist/web-extension.js",
     "watch-web": "npm run -S esbuild-web -- --sourcemap --sources-content=false --watch",
     "build": "npm run -S esbuild-base -- --sourcemap --sources-content=false && npm run -S esbuild-web -- --sourcemap --sources-content=false",
     "package": "vsce package",
@@ -43,23 +34,12 @@
     "test": "ts-mocha -p tsconfig.json tests/*test.ts --timeout 30s --diff false",
     "test:coverage": "nyc npm run test"
   },
-  "nyc": {
-    "extends": "@istanbuljs/nyc-config-typescript",
-    "check-coverage": true,
-    "all": true,
-    "include": [
-      "src/**/!(*.test.*).[tj]s?(x)"
-    ],
-    "exclude": [
-      "src/_tests_/**/*.*"
-    ],
-    "reporter": [
-      "html",
-      "lcov",
-      "text",
-      "text-summary"
-    ],
-    "report-dir": "coverage"
+  "main": "./dist/debugAdapter/index.js",
+  "dependencies": {
+    "@vscode/debugadapter": "^1.64.0",
+    "algosdk": "github:jasonpaulos/js-algorand-sdk#teal-source-map-improvements",
+    "await-notify": "^1.0.1",
+    "json-bigint": "^1.0.0"
   },
   "devDependencies": {
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
@@ -85,100 +65,22 @@
     "typescript": "^4.6.3",
     "url": "^0.11.3"
   },
-  "main": "./dist/extension.js",
-  "browser": "./dist/web-extension.js",
-  "activationEvents": [
-    "workspaceContains:**/*.teal",
-    "onDebug",
-    "onDebugResolve:avm",
-    "onDebugDynamicConfigurations:avm"
-  ],
-  "workspaceTrust": {
-    "request": "never"
-  },
-  "contributes": {
-    "languages": [
-      {
-        "id": "teal",
-        "extensions": [
-          ".teal"
-        ],
-        "configuration": "./teal-language-configuration.json"
-      }
+  "nyc": {
+    "extends": "@istanbuljs/nyc-config-typescript",
+    "check-coverage": true,
+    "all": true,
+    "include": [
+      "src/**/!(*.test.*).[tj]s?(x)"
     ],
-    "grammars": [
-      {
-        "language": "teal",
-        "scopeName": "source.teal",
-        "path": "./syntaxes/teal.tmLanguage.json"
-      }
+    "exclude": [
+      "src/_tests_/**/*.*"
     ],
-    "breakpoints": [
-      {
-        "language": "teal"
-      }
+    "reporter": [
+      "html",
+      "lcov",
+      "text",
+      "text-summary"
     ],
-    "debuggers": [
-      {
-        "type": "avm",
-        "languages": [
-          "teal"
-        ],
-        "label": "AVM Debug",
-        "program": "./out/src/debugAdapter/cli.js",
-        "runtime": "node",
-        "configurationAttributes": {
-          "launch": {
-            "properties": {
-              "simulateTraceFile": {
-                "type": "string",
-                "description": "Transaction group simulation response with execution trace.",
-                "default": "${workspaceFolder}/path/to/simulateTraceFile.json"
-              },
-              "programSourcesDescriptionFile": {
-                "type": "string",
-                "description": "Description file for sources of programs appearing in transaction group.",
-                "default": "${workspaceFolder}/path/to/programSourcesDescriptionFile.json"
-              },
-              "stopOnEntry": {
-                "type": "boolean",
-                "description": "Automatically stop after launch.",
-                "default": true
-              }
-            }
-          }
-        },
-        "initialConfigurations": [
-          {
-            "type": "avm",
-            "request": "launch",
-            "name": "Debug AVM Transactions",
-            "simulateTraceFile": "${workspaceFolder}/path/to/simulateTraceFile.json",
-            "programSourcesDescriptionFile": "${workspaceFolder}/path/to/programSourcesDescriptionFile.json",
-            "stopOnEntry": true
-          }
-        ],
-        "configurationSnippets": [
-          {
-            "label": "AVM Debug",
-            "description": "A new configuration for replaying and debugging a Algorand transactions.",
-            "body": {
-              "type": "avm",
-              "request": "launch",
-              "name": "Debug AVM Transactions",
-              "simulateTraceFile": "^\"\\${workspaceFolder}/path/to/simulateTraceFile.json\"",
-              "programSourcesDescriptionFile": "^\"\\${workspaceFolder}/path/to/programSourcesDescriptionFile.json\"",
-              "stopOnEntry": true
-            }
-          }
-        ]
-      }
-    ]
-  },
-  "dependencies": {
-    "@vscode/debugadapter": "^1.64.0",
-    "algosdk": "github:jasonpaulos/js-algorand-sdk#teal-source-map-improvements",
-    "await-notify": "^1.0.1",
-    "json-bigint": "^1.0.0"
+    "report-dir": "coverage"
   }
 }
