+++ src/cli.ts
@@ -1,8 +1,8 @@
 import * as fs from 'fs/promises';
 import { basename } from 'path';
 import * as Net from 'net';
-import { FileAccessor } from './utils';
-import { AvmDebugSession } from './debugRequestHandlers';
+import { FileAccessor } from './fileAccessor';
+import { AvmDebugSession } from './debugSession';
 
 /*
  * cli.js is the entrypoint of the debug adapter when it runs as a separate process.
