+++ src/index.ts
@@ -1,0 +1,2 @@
+export { AvmDebugSession } from './debugSession';
+export { FileAccessor } from './fileAccessor';
