+++ tsconfig.json
@@ -21,5 +21,5 @@
     "strictNullChecks": true,
     "noUnusedParameters": false
   },
-  "include": ["src/**/*", "test/**/*"]
+  "include": ["src/**/*", "extension/**/*", "tests/**/*"]
 }
